+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Tree/types.ts
@@ -2,6 +2,7 @@ import {
     type AdditionalMetric,
     type CustomDimension,
     type Dimension,
+    type GitIntegrationConfiguration,
     type GroupType,
     type Metric,
     type OrderFieldsByStrategy,
@@ -32,6 +33,8 @@ export type TreeProviderProps = {
     missingCustomMetrics?: AdditionalMetric[];
     missingCustomDimensions?: CustomDimension[];
     groupDetails?: Record<string, GroupType>;
+    isGithubIntegrationEnabled?: boolean; // For displaying the write back on custom metrics in TreeSingleNodeActions
+    gitIntegration?: GitIntegrationConfiguration;
     onItemClick: (key: string, item: NodeItem) => void;
 };